+++ rio-backend/src/config/defaults.rs
@@ -213,7 +213,7 @@ pub fn default_config_file_content() -> String {
 # shape - Default cursor shape is 'block'
 # Other available options are: 'underline', 'beam' or 'hidden'
 #
-# blinking - Whether the cursor blinks. The default is true
+# blinking - Whether the cursor blinks. The default is false
 #
 # blinking-interval - Cursor update on milliseconds interval
 #
@@ -311,9 +311,9 @@ pub fn default_config_file_content() -> String {
 # Note: You can set different font families but Rio terminal
 # will always look for regular font bounds whene
 #
-# You can also set family on root to overwrite all fonts
+# You can also set family on root to overwrite all fonts.
 # [fonts]
-#   family = "cascadiamono"
+# family = "cascadiamono"
 #
 # You can also specify extra fonts to load
 # [fonts]
@@ -325,6 +325,10 @@ pub fn default_config_file_content() -> String {
 #
 # Note: Font features do not have support to live reload on configuration,
 # so to reflect your changes, you will need to close and reopen Rio.
+#
+# You can also disable font hinting. Font hinting is enabled by default.
+# [fonts]
+# hinting = false
 #
 # Example:
 # [fonts]