+++ sugarloaf/src/sugarloaf/primitives.rs
@@ -10,6 +10,7 @@ use serde::Deserialize;
 #[derive(Debug, PartialEq, Copy, Clone)]
 pub enum SugarCursor {
     Block([f32; 4]),
+    HollowBlock([f32; 4]),
     Caret([f32; 4]),
     Underline([f32; 4]),
 }